+++ src/datajudge/requirements.py
@@ -177,7 +177,27 @@ class WithinRequirement(Requirement):
         self, column: str, condition: Condition = None, name: str = None
     ):
         ref = DataReference(self.data_source, [column], condition)
-        self._constraints.append(miscs_constraints.NullAbsence(ref, name=name))
+        self._constraints.append(
+            miscs_constraints.MaxNullFraction(ref, max_null_fraction=0, name=name)
+        )
+
+    def add_max_null_fraction_constraint(
+        self,
+        column: str,
+        max_null_fraction: float,
+        condition: Condition = None,
+        name: str = None,
+    ):
+        """Assert that ``column`` has less than a certain fraction of ``NULL`` values.
+
+        ``max_null_fraction`` is expected to lie within [0, 1].
+        """
+        ref = DataReference(self.data_source, [column], condition)
+        self._constraints.append(
+            miscs_constraints.MaxNullFraction(
+                ref, max_null_fraction=max_null_fraction, name=name
+            )
+        )
 
     def add_n_rows_equality_constraint(
         self, n_rows: int, condition: Condition = None, name: str = None
@@ -1102,6 +1122,31 @@ class BetweenRequirement(Requirement):
         self._constraints.append(
             uniques_constraints.NUniquesMaxLoss(
                 ref, ref2, max_relative_loss_getter, name=name
+            )
+        )
+
+    def add_max_null_fraction_constraint(
+        self,
+        column1: str,
+        column2: str,
+        max_relative_deviation: float,
+        condition1: Condition = None,
+        condition2: Condition = None,
+        name: str = None,
+    ):
+        """Assert that the fraction of ``NULL`` values of one is at most that of the other.
+
+        Given that ``column2``\'s underlying data has a fraction ``q`` of ``NULL`` values, the
+        ``max_relative_deviation`` parameter allows ``column1``\'s underlying data to have a
+        fraction ``(1 + max_relative_deviation) * q`` of ``NULL`` values.
+        """
+        ref = DataReference(self.data_source, [column1], condition1)
+        ref2 = DataReference(self.data_source2, [column2], condition2)
+        self._constraints.append(
+            miscs_constraints.MaxNullFraction(
+                ref,
+                ref2=ref2,
+                max_relative_deviation=max_relative_deviation,
             )
         )