+++ CHANGELOG.rst
@@ -7,11 +7,14 @@
 Changelog
 =========
 
-1.5.0 - 2022.03.XX
+
+1.5.0 - 2022.03.14
 ------------------
 
 **New features**
 
+- Implement :meth:`datajudge.BetweenRequirement.add_max_null_fraction_constraint` and
+  :meth:`datajudge.WithinRequirement.add_max_null_fraction_constraint`.
 - Implement :meth:`datajudge.BetweenRequirement.add_numeric_percentile_constraint` and
   :meth:`datajudge.WithinRequirement.add_numeric_percentile_constraint`.